build: Hook up GitHub Actions workflow to replace CodeBuild (#558)
Initially shards the lit test suite into 5 jobs in parallel, which gets the build time down to about 7.5 minutes. This can be easily tweaked later.

Also use lit functionality to skip /optimize tests on Mono, to make the build process simpler.

A separate windows build is forthcoming in the matrix, but not yet working.