feat: Handle class/trait exports more consistently
This is a breaking change. Previously, a revealed trait/class could automatically provide all its members.
Now, the signature of the trait/class (providing the name of an opaque type, or revealing the name of a
reference type) is exported separately from its members. This may require additional export-provides
or export-reveals declarations for the members.
Also, previously, the anonymous constructor of a class was exported under the name `_ctor`. Now, this
constructor is automatically exported when the class is revealed.