fix: Don't allow constructor-less `new` for imported classes
Just because an imported class appears to have no constructors (that is, the class does not export any
constructors) does not mean it has no constructors. It could be that it has a constructor that is not
exported. Therefore, the constructor-less allocation `new M.C` is not allowed for an imported class `M.C`.

The built-in types `object` and array types are known not to have any constructors, so they can be
allocated without calling a constructor. (Allocating an array in this way is most likely a user error.
Perhaps a future version of Dafny will disallow this.)

FIxed bug where cloner retained the original .Module of iterators.

Added `ClassDecl.IsObjectTrait` convenience method, which says whether or not the class is the built-in
trait `object`.