feat!: Revise export reference types (#583)
* feat: New rules for exporting classes

To _reveal_ a class or trait C will export:
- the fact that this is a reference type
- both types C and C?, and the connection between these two
- the anonymous constructor, if C is a class and has one

To _provide_ a class or trait C will export:
- the name C, as an opaque type
- not type C?
- no constructor or mutable field, and it's an error to explicitly name these in exports

Members of types are exported separately from exporting the type.
- to export a member, the enclosing class must be provided or revealed
- class constructors and mutable fields can be exported only if the enclosing class/trait is revealed
- datatype constructors, discriminators, and destructors are exported when the datatype is revealed, and can never be explicitly named in exports

Other internal implementation changes:
- Treat ClassDecl as a RevealableTypeDecl.
- Allow members in OpaqueTypeDecl (only used internally for now).
- Improve some error messages by keeping the module qualification in the name.

* fix: Improve error messages to what WhatKind of type things are

* Improve error messages by adding some quotes around identifiers

* Update various test outputs

* Add tests for aliasing of export name paths

* fix: Print `reveals *` and `provides *`

* fix: Don't allow constructor-less `new` for imported classes

Just because an imported class appears to have no constructors (that is, the class does not export any
constructors) does not mean it has no constructors. It could be that it has a constructor that is not
exported. Therefore, the constructor-less allocation `new M.C` is not allowed for an imported class `M.C`.

The built-in types `object` and array types are known not to have any constructors, so they can be
allocated without calling a constructor. (Allocating an array in this way is most likely a user error.
Perhaps a future version of Dafny will disallow this.)

* fix: FIxed bug where cloner retained the original .Module of iterators.

* chore: Added `ClassDecl.IsObjectTrait` convenience method, which says whether or not the class is the built-in trait `object`.

* Add tests that mutable fields cannot be revealed, only provided

* Make verification of the new export scheme verify

* chore: Clean up some duplicated functionality