fix: Always respect loop spec for body-less loops
Without a loop body, we want the effect that the loop specification--that
is, the loop guard, the loop invariant, and the loop frame--says all there
is to say about the loop. Dafny supplies three 3 pieces to Boogie and then
lets Boogie's loop usual processing do the rest. If the guard necessarily
holds initially and this gets noticed by Boogie, then Boogie treats this as
no loop at all. That kind of cleverness is sound, but when demonstrating
the concepts of loop reasoning, it just looks confusing. Therefore, by
this PR, Dafny counteracts this in its Boogie encoding of body-less loop

When the Dafny program supplies a loop body, then Dafny allows Boogie to
be as clever as it wishes, as before PR.