Instantiate formal type parameters of traits with their actual types when compiling
Dafny allows a type parameter of a function/method to shadow a type parameter of the enclosing type. This
does not hold true in all target languages (e.g., C#). The compiler now prepends "_" to the name of any
function/method type parameter, which removes any ambiguities.

The test cases in TraitCompile.dfy now compile correctly for C# and JavaScript.
* Compilation to Java fails, because the tests include a static method that uses a type parameter of the
  enclosing class. This is not directly supported in Java. The problem is not related to generic traits--it is
  also a problem in generic classes (that don't extend any trait).
* Compilation to Go fails. This is due to a problem in compiling generic methods.