Change type signature of CreateClass to take TopLevelDecl instead of its List<TypeParameter>.
Add originalOutParams parameter to EmitReturn.

Ignore useAllTypesArgs parameter for ...TypeDescriptors... in Java. This is refactored to use NeedsTypeDescriptor,
but at the moment, it always returns true for Java (because there's currently a dependency on the type descriptor
of sequence types in Java).

Added .Original field to Function and Method, which reflexively and transitively returns .OverriddenFunction and .OverriddenMethod.

Fix malformed Boogie code, having to do with using the correct list of type parameters.