Compile non-overridable trait members into trait companion class
Previously, these members and their implementations were duplicated into every class
that inherited them. This causes unnecessary code bloat, both the copies and from the
fact that each copy needed specialized coercions. Getting all of these coercions
correct in the compiler was also a pain and caused this part of the compiler to be
rather complicated.

The one downside with putting the non-overridable members into the trait companion
class is that it requires the compiler to know, when laying down a call to such a member,
whether or not the member is overridable. There's no syntactic indication of this in
the language, which seems against modular compilation. However, this probably already
exists with regard to whether or not an attempt of an override is legal. Therefore, this
problem is best solved at a later time.