Workaround Go embedding bug
Traits are currently compiled to a struct+interface. Unfortunately, this adds the names of parent traits
as members of an extending class, and these names may clash with the names of other members of
the class. (This could be fixed by appending, say, `_super_`, to the names of the fields that represent
the parent traits. However, such a change breaks Go's pun of embedding other structs, so the compiler
would have to insert additional `Parent_super_.` prefixes to any inherited field. This does not seem
worth it at this time, since the struct+interface encoding is about to change soon anyway.) As a
temporary workaround, this commit renames fields to avoid such clashes.