Shorten time-out in test
This change is unrelated to the fix of 867. Here's an explanation:
Z3 is slow, and evidently unpredictable, when converting between integers and bitvectors.
git-issue-505.dfy demonstrates this by providing 5 pairs of conversion tasks. Two of them
don't involve many bits and can be verified in less than 3 seconds. Each of the other tasks
typically take in excess of 60 seconds to complete. The test file specified a time-out of
20 seconds, with the expected outcome that three of the tasks would give a time-out
error. Well, with the fix to 867, something evidently perturbed the situation for Z3, so now
Z3 gave up sooner than 20 seconds, which caused the test suite to fail.

The simplest way to restore the test suite (other than removing this test altogether) was to
reduce the time-out from 20 to 8 seconds. This makes the test pass again, and also shaves
some 30 seconds off the time required to run the test.