Remove use of _ctor
Internally, the name of an anonymous class (or iterator) constructor is `_ctor`. Although this name can be used, it is not intended to be used in programs (not only is it uglier and more cryptic in the program text, but this internal name may also change to something else in the future). Thus, it also seems appropriate not to use this name in the test suite.