Remove allowNonNull parameter to Type.SameHead
With this parameter passed as “true”, the previous implementation didn’t work right. In particular, it ignored the parameter when the given type `t` had no type parameters. However, there was just one call site that used this optional parameter, so it was easier and cleaner to remove the parameter and do the necessary work at the call site.