feature: as and is for reference types (#1220)
* Reword error message to be more general

* chore: Fix code indentation

* chore: Add comment

* Remove unused parameter of `IsSupertype`

* Remove allowNonNull parameter to `Type.SameHead`

With this parameter passed as “true”, the previous implementation didn’t work right. In particular, it ignored the parameter when the given type `t` had no type parameters. However, there was just one call site that used this optional parameter, so it was easier and cleaner to remove the parameter and do the necessary work at the call site.

* Add `ignoreNullity` flag to `IsSubtypeOf` method

Note, this changes type inference slightly. There was no difference for the test suite, and I believe the change is for the better.

* Remove unnecessary `while`-`true`

* Fix typo in comment

* Tighten up code

* feat: Type check `as` for reference types

* Translate `as` for reference types

* Compile `as` for reference types

* Fill in `WhatKind` for non-null types

* Add inference of type parameters

* Add more `/titrace` output

Let `/titrace` print a debug message immediately when a constraint is flagged as an error.

* Fix typo in comment

* fix: Include type parameters in `implements$` relation

This fixes a soundness issue in the axiomatization of the class/trait implements relation.

Fixes #1212

* Test `as` cast compilation

Also, fix bug in handling types of out-parameters for multi-return calls in Java.

* Fix typo in test script

* Introduce `is` expression, `TypeTestExpr`

* Improve error-message hint for witnesses

* Make witness errors more consistent

* Type check type-test expressions

* Verification

* Fix parsing of `IsGenericInstantiation`

* Add more resolution tests

* Fix more parsing issues

* Add more verification tests

* Stub out `EmitTypeTest`

* Implement `EmitTypeTest` for C# and Java

* Fix `object` objects in Go

* Improve code

* Fix incorrect comparison in `Resolver`

* Compile `is` for JavaScript and Go

* Fix `TraitID` mechanism in Go

Go only allocates non-empty structs.

* Fix: Hand-craft `instanceof` in Go

* Fix: pay no attention to type params in Go type coercions

* Add JS and Go test answers

* Mark subset-type type tests as ghost

* Add documentation