Fix typos in Compiler-java.GenericArrayElementLvalue
This specialization of ILvalue represents an lvalue of the form `a[i]` where a is an array of unknown (that is, parameterized) type. The implementation was missing a call to `EndStmt(wr)` and had an unnecessary call to `intValue()`, and the test suite somehow managed to miss exercising this case.

This is a good case study for being able to measure the code coverage of the regression suite, which will be possible to tackle once we a) move to xUnit to drive the tests and b) stop creating separate `dafny` subprocesses for each test.