fix: typos in Compiler-java.GenericArrayElementLvalue (#1341)
This specialization of ILvalue represents an lvalue of the form `a[i]` where a is an array of unknown (that is, parameterized) type. The implementation was missing a call to `EndStmt(wr)` and had an unnecessary call to `intValue()`, and the test suite somehow managed to miss exercising this case.