Make Boogie variable names more stable (#1339)
Because Boogie has only a single scope per implementation, and Dafny can have nested lexical scopes inside a method, Dafny must rename variables when translating to Boogie to prevent name conflicts. We'd like to make the renamed variable names 'stable' in the sense that a Boogie variable name is likely to remain unchanged if unrelated parts of the Dafny program change.

Dafny uses a renaming counter to come up with fresh names when it has to rename variables. This PR creates a nested renaming counter whenever the rename algorithm steps into a Dafny scope. It's easiest to explain with an example:

```
method M() {
  var x := 0; // was and still is x#0
  assert x == 0 by {
    var x := x + 2; // used to be x#1, now is x#0_0
    var y := 0; // used to be y#0, now is y#0_0
  }
  var y := 200; // used to be y#1 because of the presence of the `var y` in the assert statement, now is y#0.
}
```

This PR adds hierarchical names where they were missing, which was in:

- Assert by statement
- The else statement of an if
- An if with a binding guard: if x :| ...
- The forall statement
- Modify statement