fix: implements$ parameters
Previously, a declaration

``` dafny
trait T<A, B> { ... }
```

would generate a function

```
function implements$T(T$A: Ty, T$B: Ty, Ty) : bool;
```

in Boogie. When printed, this won't parse back in as Boogie, because Boogie says that either all or none of the parameters of a function should be given names.
The names aren't used for anything, other than as a guide to a human reader. But as such, the parameters are also in the wrong order.

This commit fixes both of these problems to instead generate

```
function implements$T(ty: Ty, T$A: Ty, T$B: Ty) : bool;
```