Translator breakup (#1337)
This PR is a bag of refactor commits that mostly move code. Feel free to disagree with this change or with some of the individual commits. I hope this PR can at least start some discussion of how we want to structure the code and how to get there.

I found it difficult to find structure in Translator.cs. This PR tries to add structure by moving related code into separate files. To make the move relatively simple, the change uses C#'s partial classes so all the code moved out of Translator.cs is still in the Translator class. About 5k lines of codes are moved out of Translator.cs changing it from 19k to 14k lines.

The separate files are:

Translator.ExpressionTranslator.cs, contains code for translating expressions
Translator.Statement.cs, contains code for translating statements.
Translator.BoogieFactory.cs, contains code that's independent of Dafny, that helps with generating Boogie