Show a source code snippet for each Dafny message (#1321)
This commit adds a /showSnippets argument to shows a source code snippet for each Dafny message, similar to what the Rust compiler does.

This is disabled by default to be backward compatible with whatever tool might be parsing the output. If it's not an issue, however, I think it would be useful to enable this by default.
The reported snippet is not guaranteed to be the exact same code that Dafny parsed, because the file might have been modified on disk in the meantime.