Language Server: Fix diagnostics for windows (#1345)
A recent push has introduced issues on windows relating to the paths. Unfortunately, cross-platform tests were removed to simplify the builds when merging into the main repo; thus, it wasn't captured earlier.
This PR fixes the path issues on Windows. Furthermore, the PR does some minor refactorings to centralize all path-related operations in a single class PathExtensions to ensure consistency in the future.

The matrix build is added again to test against ubuntu/windows/osx in the future.

Co-authored-by: Remy Willems <rgv.willems@gmail.com>