Make building Dafny against a custom Boogie easier (#1420)
Fixes #1310

This PR adds a customBoogie.patch file to the Dafny repository, to help with building against a local Boogie. You can apply the patch with git apply customBoogie.patch and remove it with git checkout .

The INSTALL wiki has been updated to incorporate this change: https://github.com/dafny-lang/dafny/wiki/INSTALL#build-against-a-custom-boogie.

I looked at alternatives for building against a custom Boogie, with a number mentioned here. An interesting option was dnb, which has a command dnb switch-to-projects that allows switching NuGet references to project references, but sadly that tool didn't work for me.