feat: Ghost allocations (#1443)
## Summary

Previously, Dafny outlawed allocations (that is, using `new`) in ghost contexts. This PR removes that restriction. The PR also allow a `ghost method` (but not lemmas) to allocate state.

## Illustrations and more details

For example, the statements

``` dafny
ghost var a := new int[20];
ghost var b := new int[] [2, 3, 5, 7, 11, 13];
ghost var c := new int[100](i => 2 * i);
```

are now allowed, and so are the assignments in

``` dafny
class Cell {
  ghost var gdata: int
  ghost method GInit(x: int)
    modifies this
  {
    gdata := x;
  }
}

method Example0() {
  ghost var c: Cell;
  c := new Cell;
  c := new Cell.GInit(12);
}
```

and also those in

``` dafny
class XY {
  var x: int
  ghost var y: int

  constructor Init(n: int) {
    x, y := n, n;
  }

  ghost constructor GInit(n: int) {
    x, y := n, n;
  }
}

method Example1() {
  var a := new XY.Init(10);
  ghost var b := new XY.GInit(10);
}
```

Since the language gives no way to "leak" the value of a ghost variable into a non-ghost variable, an object or array allocated in a ghost context will forever remain in a ghost context. Consequently, after initialization, only ghost fields of such an object can ever be modified. Note that array elements are always non-ghost, so an array allocated in a ghost context is in effect read-only. The initialization of an array (in particular, the initializers given) or object (in particular, the initialization phase of a ghost constructor) can give initial values to non-ghost state.

Auto-ghost declarations are handled in the usual manner. For example, if `F` is a ghost function, then

``` dafny
var arr := int[12](F);
```

automatically makes `arr` a ghost variable.

As mentioned above, this PR also changes the language to allow a `ghost method` to use `new`. However, a lemma (`lemma`, `twostate lemma`, `least lemma`, `greatest lemma`) is still not allowed to use `new`.