fix: Set/forall comprehensions with subtyping for issue #697 (#1522)
This PR is a quick fix for issues #697 and #698.

The problem: Unsound comprehensions with subset types.

For-comprehensions, Exist-comprehensions, Map-comprehensions and Set-comprehensions had a problem when encountering subset types: Dafny was not emiting the code of the subset constraint, to ensure that the constraint of the comprehension is not executed with undesirable data.
The solution implemented in this PR:
Added field [ConstraintIsCompilable](https://github.com/dafny-lang/dafny/pull/1522/files#diff-7d845d452486b0b59700ca5c133f2648f76d44616148081d51efffe32dbc6e7aR2680) for SubsetTypeDecl expressions.
If a comprehension uses a subsettype whose ConstraintIsCompilable is false, will raise two errors, the second one indicating why it is not compilable
In the compiler, the function [MaybeInjectSubsetConstraint](https://github.com/dafny-lang/dafny/pull/1522/files#diff-998a9162635a910ca366a26c8f679ea96d6bbb3f703a52ca1487fd19599de49dR4743) which injects the guard code for for-comprehensions, exists-comprehensions, map-comprehensions and set-comprehensions.

How I tested it.

I added 8 tests:

git-issues/git-issue-697.dfy: A set comprehension with compilable subset constraint that did not compile correctly and used to crash at execution. Now it compiles and does not crash.
git-issues/git-issue-697b.dfy: A map comprehension with compilable subset constraint that did not compile correctly and used to crash at execution. Now it compiles and does not crash.
git-issues/git-issue-697c.dfy: A set comprehension, a map comprehension and a forall comprehension with ghost subset constraint that used to compile and crash at execution. Now it does not compiles and raises six errors.
git-issues/git-issue-697d.dfy: A set comprehension with ghost & compilable subset constraint that used to compile and crash at execution. Now it compiles and does not crash.
git-issues/git-issue-697e.dfy: A forall comprehension with only inferred ghost subset constraint. It used to compile and crash at execution time. Now it compiles and does not crash.
git-issues/git-issue-697f.dfy: A ghost subset constraint that is never used in comprehensions. It used to compile and continues to compile, ensuring this PR does not break this existing desirable behavior.
git-issues/git-issue-698.dfy: A forall comprehension with compilable subset constraint that did not compile correctly and used to crash at execution. Now it compiles and does not crash.
git-issues/git-issue-698b.dfy: An exist comprehension with compilable subset constraint that did not compile correctly and used to crash at execution. Now it compiles and does not crash.