feat(lsp): Report tracing messages to language server clients
This is intended to expose to LSP clients a feature that already exists in
Emacs' dafny-mode (coarse-grained progress reporting).

In the long term the proper way to do this is to have Dafny report fine-grained
progress as it runs verification tasks, but for now we can leverage the
existing API to display simple progress messages.

* Workspace/ICompilationStatusNotificationPublisher.cs,
Workspace/CompilationStatusNotificationPublisher.cs (SendStatusNotification):
Accept an additional parameter for progress messages.

* Language/IProgramVerifier.cs, Language/DafnyProgramVerifier.cs (Verify):
Accept a IVerificationProgressReporter instance.
* Language/DafnyProgramVerifier.cs (ModelCapturingOutputPrinter): Forward
Inform() messages to it.

* Language/IVerificationProgressReporter.cs: New interface to report
verification progress.
* Workspace/TextDocumentLoader.cs (VerificationProgressReporter): Implement it
by forwarding notifications to an ICompilationStatusNotificationPublisher.