fix: Don't get stuck when onsave verification cancelled via LSP
Previously, if an LSP client triggered the verification of a file because
of a document save while a previous verification was still running, the
process of cancelling the first verification was badly behaved. The thread
handling the verification would throw an exception and the client would
not get a response. This would lead to the VSCode client getting stuck
displaying "Verifying..." forever.

This commit fixes the issue with two changes:

* It adds a catch block in the handler for a save event that behaves
  similarly to that for a change event in the case where verification
  is cancelled. This block surrounds the code that adds a new verification
  task to the queue of operations to be performed in the future.
* It changes the body of the catch block surrounding execution of document
  verification to ensure that it responds to the client even if
  verification is cancelled.