feat: Log verification of splits (#1656)
For now this is only active when /vcsSplitOnEveryAssert is on, and replaces the per-method test results instead of augmenting them. I need to figure out if the TRX model supports sub-results well to do this better.

Also add support for `/verificationLogger:trx;LogFileName=...` so the test can be deterministic and not leak TRX files everywhere. :)

Also implement %OutputCheck in order to have a real test for the TRX generation process!