fix(lsp): Fix a concurrency bug in `AddDiagnosticForFile` (Fixes #1651)
Sending a didOpen message to the server produces roughly the following call
tree:

    └┬─DocumentDatabase.OpenDocumentAsync
     └┬─DocumentDatabase.OpenAndVerifyAsync
      └┬─DocumentDatabase.LoadAndVerifyAsync
       └┬─TextDocumentLoader.VerifyAsync
        └┬─TextDocumentLoader.VerifyInternal
         └┬─DafnyProgramVerifier.Verify
          ├┬─DafnyProgramVerifier.VerifyWithBoogie
          │└┬─ExecutionEngine.InferAndVerify
          │ └┬─ExecutionEngine.VerifyImplementation
          │  └─ ExecutionEngine.ProcessErrors
          └┬─(Callback) ModelCapturingOutputPrinter.WriteErrorInformation
           └┬─DiagnosticErrorReporter.ReportBoogieError
            └─ DiagnosticErrorReporter.AddDiagnosticForFile

There can be multiple concurrent calls to `AddDiagnosticForFile`, but the code
doesn't protect the datastructures that this function touches from concurrent
modifications.  Hence we can observe traces like this:

    [lsp] Verifying t26: Enter WriteErrorInformation
    [lsp] Verifying t27: Enter WriteErrorInformation
    [lsp] Verifying t25: Enter WriteErrorInformation
    [lsp] Verifying t25: Exit WriteErrorInformation
    [lsp] Verifying t26: Exit WriteErrorInformation
    [lsp] Verifying t27: Exit WriteErrorInformation

Which in turn cause issues with the call to `diagnostics.GetOrCreate`, since we
are in `didOpen`:

    diagnostics.GetOrCreate(documentUri, () => new List<Diagnostic>());

There's probably also an issue with the increment to count, since it's not
atomic:

    counts.TryGetValue(severity, out var count);
    counts[severity] = count + 1;

In fact, even the call to `fileDiagnostics.Add(item)` is probably unsafe, since
that's not a concurrent collection.

* DiagnosticErrorReporter.cs (AddDiagnosticForFile): Wrap function body in a
lock to avoid concurrent updates.