fix(lsp): Don't get stuck when onsave verification is cancelled (#1652)
* fix: Don't get stuck when onsave verification cancelled via LSP

Previously, if an LSP client triggered the verification of a file because of a document save while a previous verification was still running, the process of cancelling the first verification was badly behaved. The thread handling the verification would throw an exception and the client would not get a response. This would lead to the VSCode client getting stuck displaying "Verifying..." forever.

This commit fixes the issue with several changes:

  * It adds a catch block in the handler for a save event that behaves similarly to that for a change event in the case where verification is cancelled.
  * It refactors SaveDocumentAsync to be more similar to UpdateDocumentAsync, to ease understanding.
  * It makes the check for in-order change messages more liberal, allowing non-consecutive version numbers (as allowed by the LSP standard, and as occur in practice after save events).
  * It includes a test that fails with previous code but succeeds after these changes.