docs: add material on interacting with java code (#1686)
Add two examples about passing strings between Dafny and Java code. In one example, the extern method obtains a Java string and returns it as a Dafny string. In the other, the extern method obtains a Java string and returns it as such. The Dafny code converts that Java string to a Dafny string.