Update use-local-boogie.sh to .NET 6.0
Also make it use Release configuration builds. Given that it copies
binaries already, it seems that it makes sense to get optimized
binaries. If you want to debug Boogie along with Dafny, the
customBoogie.patch approach seems more appropriate, as it makes the
source more readily available in the IDE.