feat: `DowncastClone` for functions and datatypes with contravariant type parameters (#1755)
This is a follow-up to #1645, which omitted to generate DowncastCopy for datatypes with contravariant type parameters. To make this useful we also add it for functions, but still don't generate DowncastCopy, when a datatype has a constructor expecting an argument of a type that contains a generic reference type reusing the datatype's type parameters.