Trace through match expressions for more precise errors (#1768)
This PR lets "related location" messages trace through the cases of match expressions. The implementation change itself is small, but it affects the generation of Boogie code--in particular, it affects `TrSplitExpr`, which is called for every proof obligation generated from the user-provided expressions. Thus, the improvements in this PR may be destabilizing for some verifications.

The PR also contains a couple of bug fixes, which were discovered during the testing of the new feature and which were needed in order for the new feature to work. Finally, the PR also contains a few code beautifications.

* Improve error message, because "cannot safely" sounds like a user error

* Handle "match" in TrSplitExpr

* Apply induction only in Impl specs, not Call and WellFormed specs

* Don't exclude TrSplitExpr-inlining function with "match" body

* fix: Let CanSafelyInline look at desugared body of prefix predicates

* Remove unnecessary semi-colons

* Increase error limit to show all errors

* refactor: Factor TrSplitFunctionCallExpr

Fixes #977