Support for Mocking in C# Compiler (#1809)
### UPDATE (Mar 28 2022):

As per the ensuing discussion, the functionality of the `{:mock}` and `{:fresh}` attributes has been merged into a single `{:synthesize}` attribute. Methods annotated with `{:synthesize}`  no longer need to be annotated with `{:extern}`. The compilation behavior described below no longer requires a special CLI option to be triggered. Finally, the compiler will now throw an error if a stubbed method includes any post-conditions, which should guard against a use of the `{:synthesize}` attribute that might introduce unsoundness.

### ORIGINAL PR DESCRIPTION:

This PR adds mocking support to the C# compiler. The idea is to have a special annotation for external methods that makes the compiler translate a postcondition to a series of API calls in C#'s Moq library (provided the `/compileMocks` command-line argument is also supplied). As an example, consider the following Dafny code, which tests the static method `getOrDefault` of class `StringMap` by mocking a `StringMap` object and stubbing its two instance methods. Using this approach, one is able to test the implementation of `getOrDefault` without specifying how one would add elements to the map.

```Dafny
class StringMap {

	var m:map<string, string>;

	function method Contains(key:string):bool reads this {
		key in m
	}

	function method Get(key:string):string reads this
		requires Contains(key)
	{
		m[key]
	}

	static function method GetOrDefault(m:StringMap, key:string, default:string):string reads m
	{
		if m.Contains(key) then m.Get(key) else default
	}
}

method {:extern} {:mock} MockStringMap(k:string, v:string) 
	returns (m:StringMap)
	ensures m.Contains(k) == true && m.Get(k) == v
	ensures fresh(m)

method {:test} PassingTestGetOrDefault() {
	var stringMap := MockStringMap("a", "b");
	assert StringMap.GetOrDefault(stringMap, "a", "c") == "b";
	expect StringMap.GetOrDefault(stringMap, "a", "c") == "b";
}
```

The `mockStringMap` method is automatically translated to the following in C#, allowing a runtime test that would otherwise be impossible without fully implementing the StringMap class:

```cs
public static StringMap mockStringMap(Dafny.ISequence<char> k, Dafny.ISequence<char> v) {
    var mTmp = new Mock<StringMap>();
    mTmp.Setup(x => x.contains(k)).Returns(true);
    mTmp.Setup(x => x.get(k)).Returns(v);
    return mTmp.Object;
}
```

Only functions can be mocked - methods can have side effects and are, therefore, not allowed in postconditions. Not all Dafny postconditions can be translated to calls in Moq either - compiling an arbitrary postcondition would be a full-blown synthesis problem and, while intersting from a research perspective, it would probably not be fast enough for a compiler. Here is the grammar for postconditions that are supported (`S` is the start symbol, `EXP` stands for an arbitrary Dafny expression, and `ID` stands for variable/method/type identifiers):

```
S       = FORALL 
        | EQUALS 
        | S && S
EQUALS  = ID.ID (ARGLIST) == EXP // stubs a method call
        | ID.ID            == EXP // stubs field access
        | EQUALS && EQUALS
FORALL  = forall SKOLEMS :: EXP ==> EQUALS
ARGLIST = ID // this can be one of the skolem variables
        | EXP // this exp may not reference any of the skolems 
        | ARGLIST, ARGLIST
SKOLEMS = ID : ID 
        | SKOLEMS, SKOLEMS
```

The compiler throws a NotImplementedException if postconditions on the mocked method do not follow this grammar. Support for mocking could be easily extended to Java and other target languages (and I am happy to work on this). Here is a [design document](https://github.com/Dargones/TestingFrameworkPaper/blob/master/main.pdf) that outlines the overall idea of a testing framework and of how one would do this in Java.

An additional feature that this PR introduces is the `{:fresh}` annotation that prompts the compiler to translate an external method to one that returns a fresh instance of an object. This could be useful in situations where inferring the right parameters to a constructor is difficult. So, for instance, an automatic test generation framework could use this to get a fresh instance of an object to then assign its fields as needed. This is the feature that was previously the topic of [this PR](https://github.com/dafny-lang/dafny/pull/1526).

Finally, this PR also fixes a small issue with the `{:test}` annotation - currently, the C# compiler only calls `IsFailure` on the first returned value of a testing method if there are more than one (should be more than zero) returned value ([this](https://github.com/dafny-lang/dafny/blob/c6e84137226510094fd9079534e967d35f07a3f3/Source/Dafny/Compilers/Compiler-Csharp.cs#L3185) is the line with the bug)